refactor utils files usage and delete unneded tests
