--- deeptracy/utils.py
@@ -1,194 +1,0 @@
-# Copyright 2017 BBVA
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import requests
-import os
-import yaml
-import uuid
-import docker
-import logging
-
-from shutil import copyfile
-
-import deeptracy.config as config
-from deeptracy_core.dal.project.model import RepoAuthType
-
-logger = logging.getLogger(__name__)
-logger.setLevel(logging.DEBUG)
-
-
-def valid_repo(url: str):
-    """Given a repository url checks if the repository is publicly accessible
-
-    Args:
-        url (str): Repository URL. Must be a full url, including schema.
-
-    Returns:
-        True for success, False otherwise
-
-    Raises:
-        requests.exceptions.RequestException
-    """
-    if url is None:
-        raise ValueError('Missing repo url')
-
-    # normalize repo url
-    url = url.strip().replace(" ", "")
-    # Check if remote is accessible
-    if requests.get(url, timeout=20).status_code != 200:
-        return False
-
-    return True
-
-
-def prepare_path_to_clone_with_local_key(scan_path: str, repo: str, mounted_vol: str, source_folder: str):
-    """
-    Prepare a folder to clone a repository which needs a local private key.
-
-    LOCAL_PRIVATE_KEY means we need to copy the local private key present in the host to the folder
-    that is being mounted in to the container that is going to perform the actual repo clone
-
-    We prepare a script with all the commands needed to perform the clone, like adding the private key
-    to the container ssh-agent and disabling host-key verification
-
-    :param scan_path: (str) path that we need to prepare
-    :param repo: (str) project repository to clone
-    :param mounted_vol: (str) path to the mounted volume in the container that makes the clone
-    :param source_folder: (str) name of the folder to made the actual clone
-    :return: returns the command to pass to the container that makes the clone
-    """
-    key_name = 'local_id_rsa'
-    dest_key_path = os.path.join(scan_path, key_name)
-    copyfile(config.LOCAL_PRIVATE_KEY_FILE, dest_key_path)
-
-    # Having the key in the mounted folder, preapre a script to activate the ssh-agent, disable the host
-    # key verification and add the key to it after doing the clone
-    script_contents = '#!/bin/bash \n' \
-                      'eval `ssh-agent -s` \n' \
-                      'chmod 400 {mounted_vol}/{key} \n' \
-                      'mkdir -p /root/.ssh \n' \
-                      'touch /root/.ssh/config \n' \
-                      'echo -e "StrictHostKeyChecking no" >> /root/.ssh/config \n' \
-                      'ssh-add {mounted_vol}/{key} \n' \
-                      'git clone {repo} {mounted_vol}/{source_folder}\n' \
-        .format(
-            key=key_name,
-            repo=repo,
-            mounted_vol=mounted_vol,
-            source_folder=source_folder
-        )
-
-    # create the script that makes the clone
-    script = os.path.join(scan_path, 'clone.sh')
-    with open(script, "w") as f:
-        f.write(script_contents)
-    os.system('chmod +x {}'.format(script))
-
-    command = os.path.join(mounted_vol, 'clone.sh')
-    return command
-
-
-def clone_project(base_path: str, scan_id: str, repo_url: str, repo_auth_type: str) -> str:
-    """
-    Clone a project repository.
-
-    This method handles repository auth if the repo is not public. The repository is going to be cloned
-    in "(base_path)/(scan_id)/sources" folder (it will be created).
-
-    To do the clone a docker image is used: `bravissimolabs/alpine-git`
-
-    :param base_path: (str) Base path to clone. Should be an absolute path
-    :param scan_id: (str) Scan id that triggers the clone. Its going to be created as a folder inside the base_path
-    :param repo_url: (str) project repository url to make the git clone
-    :param repo_auth_type: (str) if the repo needs any kind of auth
-
-
-    :return: returns the sources path where the repo is cloned
-    """
-    source_folder = 'sources'
-    scan_path = os.path.join(base_path, scan_id)
-    scan_path_sources = os.path.join(scan_path, source_folder)
-    if not os.path.exists(scan_path):
-        os.makedirs(scan_path)
-
-    mounted_vol = '/opt/deeptracy'
-
-    # Command to run for public repos
-    command = 'git clone {repo} {mounted_vol}/{source_folder}/'.format(
-        repo=repo_url,
-        mounted_vol=mounted_vol,
-        source_folder=source_folder
-    )
-
-    if repo_auth_type == RepoAuthType.LOCAL_PRIVATE_KEY.name:
-        # if the project is auth with LOCAL_PRIVATE_KEY prepare the path and get the new command to clone
-        command = prepare_path_to_clone_with_local_key(scan_path, repo_url, mounted_vol, source_folder)
-
-    logger.debug('clone repo with command {}'.format(command))
-
-    docker_client = docker.from_env()
-
-    # prepare mounted volumes
-    docker_volumes = {
-        scan_path: {
-            'bind': mounted_vol,
-            'mode': 'rw'
-        }
-    }
-
-    # if utils.valid_repo(project.repo) is False:
-    #     update_scan_state(scan, ScanState.INVALID_REPO, session)
-    #     return
-
-    # launch the container with a command that will clone the repo
-    docker_client.containers.run(
-        image='bravissimolabs/alpine-git',
-        command=command,
-        remove=True,
-        volumes=docker_volumes
-    )
-
-    logger.debug('repo {} cloned'.format(repo_url))
-
-    return scan_path_sources
-
-
-def parse_deeptracy_yml(source_dir: str):
-    """
-    Find a .deeptracy.yml file inside source_dir and try to parse it.
-
-    If the file is not found, return None
-
-    :param source_dir:
-    :return: None is the file is not found or cant be parsed, else it returns a dict with the key/values
-    """
-    logger.debug('try parse .deeptracy.yml at {}'.format(source_dir))
-    yml_file = os.path.join(source_dir, '.deeptracy.yml')
-    if os.path.isfile(yml_file):
-        logger.info('.deeptracy.yml FOUND in {}'.format(source_dir))
-
-        with open(yml_file, 'r') as contents:
-            parsed = yaml.load(contents)
-
-        assert type(parsed) is dict  # should be parsed as a dict
-        assert 'lang' in parsed  # should have a lang attr
-        assert type(parsed.get('lang')) is str  # TODO: assert against a valid lang ENUM
-
-        return parsed
-
-    return None
-
-
-def make_uuid() -> str:
-    return uuid.uuid4().hex
