--- tests/unit/base_test.py
@@ -1,31 +1,0 @@
-# Copyright 2017 BBVA
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import unittest
-import os
-
-
-class BaseDeeptracyTest(unittest.TestCase):
-    """Base test class for tests suits
-
-    Its ensure a valid environment is set before running
-    each test."""
-
-    def setUp(self):
-        """Setup before each test is run"""
-        # setup environment
-        os.environ["REDIS_URI"] = "redis://localhost"
-        os.environ["POSTGRES_URI"] = "sqlite:///:memory:"
-        # load config
-
